requirements.txt
# Core dependencies
fastapi==0.68.1
uvicorn==0.15.0
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.5
pydantic==1.8.2
sqlalchemy==1.4.23
psycopg2-binary==2.9.1
python-dotenv==0.19.0

# Security and monitoring
prometheus-client==0.11.0
opentelemetry-api>=1.7.1
opentelemetry-sdk>=1.7.1
opentelemetry-instrumentation-fastapi>=0.24b0
opentelemetry-exporter-otlp>=1.7.1

# Machine Learning
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.0.0
joblib>=1.0.0
tensorflow>=2.6.0
keras>=2.6.0

# API and Integration
requests>=2.26.0
aiohttp>=3.7.4
websockets>=9.1
python-socketio>=5.4.0

# Testing and Development
pytest>=6.2.5
pytest-asyncio>=0.15.1
pytest-cov>=2.12.1
black>=21.7b0
flake8>=3.9.2
mypy>=0.910

# Documentation
mkdocs>=1.2.3
mkdocs-material>=7.2.6
mkdocstrings>=0.15.2

# Additional dependencies
boto3>=1.28.38
python-telegram-bot>=20.4
discord-webhook>=1.3.0 